# server/swagger.yaml
openapi: 3.0.0
info:
  title: Kanata Scraper API
  description: API documentation for Kanata Scraper (Otakudesu, etc.)
  version: 1.0.0
servers:
  - url: http://localhost:3000
    description: Local Development Server
paths:
...
  /api/scraper/tiktok:
    get:
      tags:
        - Downloader
      summary: TikTok Downloader
      description: Downloads TikTok video without watermark.
      parameters:
        - in: query
          name: url
          schema:
            type: string
          required: true
          description: TikTok video URL
      responses:
        '200':
          description: Video data
          content:
            application/json:
              schema:
                type: object
                properties:
                  title: { type: string }
                  author: { type: string }
                  cover: { type: string }
                  nowm: { type: string }
                  music: { type: string }
                  stats: { type: object }
...
  /api/scraper/sholat/{cityId}:
    get:
      tags:
        - Islamic
      summary: Get Prayer Schedule
      description: Returns daily prayer schedule for a specific city ID.
      parameters:
        - in: path
          name: cityId
          schema:
            type: string
          required: true
          description: City ID from the /cities endpoint
      responses:
        '200':
          description: Prayer schedule
          content:
            application/json:
              schema:
                type: object
...